2.0
The content of this update is as follows:
Modified part of the multi-process structure, reducing the memory usage to about 2GB
Up to 8 threads are supported.
Added avx optimization option.
Added command line run mode.
Now, it takes 45 minutes (in 8 threads) for this program to process 10GB of data. Using avx will cut the time in half.
The program may exit abnormally when running with avx.